JENA-714 : Language upgrade
git-svn-id: https://svn.apache.org/repos/asf/jena/trunk@1605034 13f79535-47bb-0310-9956-ffa450edef68

+++ jena-core/src/main/java/com/hp/hpl/jena/reasoner/rulesys/impl/RETEQueue.java
@@ -68,7 +68,7 @@ public class RETEQueue implements RETESinkNode, RETESourceNode {
 		int len = matchIndexList.size();
 		matchIndices = new byte[len];
 		for (int i = 0; i < len; i++) {
-			matchIndices[i] = matchIndexList.get(i).byteValue();
+			matchIndices[i] = matchIndexList.get( i );
 		}
 		this.queue = new BindingVectorMultiSet(matchIndices);
 	}
